requirements.txt
SQLAlchemy
fastapi
uvicorn
starlette
pydantic
passlib
python-multipart
requests
pydantic[email]
python-jose[cryptography]